requirements.txt
requests
beautifulsoup4
python-slugify
html_to_markdown
openai
tqdm